services/rag_chatbot_service/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic
python-dotenv
python-multipart>=0.0.6
requests

# Database
sqlalchemy>=1.4.0

# AI/LLM
google-generativeai